fleet-backend/requirements.txt
# Flask Backend - Fleet Management System
# Python 3.11+

# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# Database
pymysql==1.1.0  # MySQL
cryptography==41.0.7  # Para pymysql
SQLAlchemy==2.0.23

# Async Tasks
celery==5.3.4
redis==5.0.1
flower==2.0.1  # Celery monitoring

# HTTP Requests
requests==2.31.0
zeep==4.2.1  # SOAP client para Ituran

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2

# Monitoring & Logging
sentry-sdk[flask]==1.39.1

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
